requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.5
django-cors-headers==4.6.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
urllib3==2.3.0
